requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
psycopg[binary]>=3.1
python-docx>=0.8
chromadb>=0.4
openai>=1.12
pydantic>=2.6
